requirements.txt
moviepy==2.2.1
faster-whisper==1.2.1
requests==2.33.1
python-dotenv==1.2.2
google-api-python-client==2.193.0
piper-tts==1.4.2
PyYAML==6.0.3
tavily-python==0.7.23
gspread==6.2.1
google-auth==2.49.1
google-auth-oauthlib==1.3.1
google-auth-httplib2==0.3.1
